+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog laravel-repository
 
+## 1.2.0 - 2018-06-14
+
+- Add method `scopes()`
+- Add method `firstOrFail()`
+- Update pagination: add `request('per_page')`
+
 ## 1.1.0 - 2018-01-13
 
 - Add cache, config